Vision/LocalSurfaceEmbeddings/requirements.txt
absl-py==1.0.0
addict==2.4.0
anyio==3.5.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
Babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.11.1
black==19.10b0
bleach==4.1.0
brotlipy==0.7.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==37.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.4
entrypoints==0.3
etils==0.5.1
executing==0.8.3
fastjsonschema==2.15.3
flatbuffers==2.0
fonttools==4.25.0
freetype-py==2.3.0
gast==0.5.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.1
grpcio==1.46.1
h5py==3.6.0
hsluv==5.0.2
idna==3.3
imageio==2.19.2
importlib-metadata==4.11.3
importlib-resources==5.7.1
ipykernel==6.9.1
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.7
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyter-packaging==0.12.0
jupyter-server==1.17.0
jupyterlab==3.4.0
jupyterlab-pygments==0.1.2
jupyterlab-server==2.13.0
jupyterlab-widgets==1.0.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==14.0.1
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.2
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
mypy-extensions==0.4.3
nbclassic==0.3.7
nbclient==0.5.11
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.1
notebook==6.4.8
notebook-shim==0.1.0
numpy==1.22.3
oauthlib==3.2.0
open3d==0.15.2
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip==22.1.1
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.20.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyglet==1.5.26
Pygments==2.11.2
PyOpenGL==3.1.6
PyOpenGL-accelerate==3.1.5
pyOpenSSL==22.0.0
pyparsing==3.0.4
pyquaternion==0.9.9
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.0
setuptools==62.1.0
sip==4.19.13
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.2.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
tensorflow-metadata==1.8.0
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
tf-estimator-nightly==2.8.0.dev2021122109
tfds-nightly==4.5.2.dev202205140045
threadpoolctl==2.2.0
tinycss2==1.1.1
toml==0.10.2
tomlkit==0.10.2
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typed-ast==1.4.3
typing_extensions==4.1.1
urllib3==1.26.9
vispy==0.10.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.1.2
wheel==0.37.1
widgetsnbextension==3.5.2
wrapt==1.14.1
zipp==3.7.0
